Fix activesupport gem name
The proper name of the Active Support gem seems to be `activesupport`. There is an `active_support` gem, but it has been yanked. This doesn't change the paths used with `require` which still include `active_support`.

We cannot upgrade to `breach-mitigation-rails` without this change.

Note: I edited the gemspec on GitHub, so I haven't actually verified that this works, but I have confirmed the proper gem name.